TOY-39 MainActivity 제거
1. MainActivity 클래스를 제거했습니다.

2. CameraActivity 코드를 CameraFragment로 옮겼습니다.

3. CameraActivity에 대해 권한 설정 조작 등 OS에 의해 종료되면 onCreate()부터 다시 시작하도록 수정했습니다.

4. CameraViewModel의 context를 global application 의 context로 대체했습니다.

5. xml의 tools:context가 MainActivity인 코드를 제거했습니다.

6. strings에서 MainActivity를 CameraFragment로 수정했습니다.
