--- app/src/main/java/kr/co/yogiyo/rookiephotoapp/MainActivity.java
@@ -1,55 +1,0 @@
-package kr.co.yogiyo.rookiephotoapp;
-
-import android.Manifest;
-import android.content.Intent;
-import android.os.Build;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.gun0912.tedpermission.PermissionListener;
-import com.gun0912.tedpermission.TedPermission;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import kr.co.yogiyo.rookiephotoapp.ui.camera.capture.CameraActivity;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        setupCheckPermission();
-    }
-
-    private void setupCheckPermission() {
-        PermissionListener permissionListener = new PermissionListener() {
-            @Override
-            public void onPermissionGranted() {
-                startActivity(new Intent(MainActivity.this, CameraActivity.class));
-            }
-
-            @Override
-            public void onPermissionDenied(ArrayList<String> deniedPermissions) {
-                finish();
-            }
-        };
-
-        List<String> permissions = new ArrayList<>();
-        permissions.add(Manifest.permission.CAMERA);
-        permissions.add(Manifest.permission.WRITE_EXTERNAL_STORAGE);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            permissions.add(Manifest.permission.READ_EXTERNAL_STORAGE);
-        }
-
-        TedPermission.with(this)
-                .setPermissionListener(permissionListener)
-                .setDeniedMessage(R.string.text_denied_permission)
-                .setGotoSettingButtonText(R.string.text_settings)
-                .setDeniedCloseButtonText(R.string.text_cancel)
-                .setPermissions(permissions.toArray(new String[0]))
-                .check();
-    }
-}
