--- app/src/main/res/layout/activity_main.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:id="@+id/root_frame"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-</android.support.constraint.ConstraintLayout>
